# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/specification": {
            "get": {
                "description": "Get a list of Specification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Get all Specification.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Specification"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Specification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Create New Specification.",
                "parameters": [
                    {
                        "description": "the body to create a new Specification",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.specificationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Specification"
                        }
                    }
                }
            }
        },
        "/specification/{id}": {
            "get": {
                "description": "Get an Specification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Get Specification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Specification"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Specification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Delete one Specification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Specification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Update Specification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age specification category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.specificationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Specification"
                        }
                    }
                }
            }
        },
        "/specification/{id}/devices": {
            "get": {
                "description": "Get all Devices by SpecificationId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Get Devices.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Devices"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.specificationInput": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "camera": {
                    "type": "string"
                },
                "capacity_battery": {
                    "type": "string"
                },
                "chipset": {
                    "type": "string"
                },
                "launched": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "ram": {
                    "type": "string"
                },
                "resolution_screen": {
                    "type": "string"
                },
                "size_screen": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                },
                "touch_screen": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "models.Devices": {
            "type": "object",
            "properties": {
                "brandID": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "specificationID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Specification": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "camera": {
                    "type": "string"
                },
                "capacity_battery": {
                    "type": "string"
                },
                "chipset": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "launched": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "ram": {
                    "type": "string"
                },
                "resolution_screen": {
                    "type": "string"
                },
                "size_screen": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                },
                "touch_screen": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        }
    }
}